+++ R/rtrunc_direct.R
@@ -82,6 +82,26 @@ rtrunc_direct.invgamma <- function(
   return(q_T)
 }
 
+#' @export
+rtrunc_direct.invgauss <- function(
+  n, family, m, s, a = 0, b = Inf, ...
+) {
+  F_a <- cumDens(a, pinvgauss, m, s)
+  F_b <- cumDens(b, pinvgauss, m, s)
+  q_T <- truncated_q(qinvgauss(rescaled_q(n, F_a, F_b), m, s), mget(ls()))
+  return(q_T)
+}
+
+#' @export
+rtrunc_direct.lognormal <- function(
+  n, family, meanlog, sdlog, a = 0, b = Inf, ...
+) {
+  F_a <- cumDens(a, plnorm, meanlog, sdlog)
+  F_b <- cumDens(b, plnorm, meanlog, sdlog)
+  q_T <- truncated_q(qlnorm(rescaled_q(n, F_a, F_b), meanlog, sdlog), mget(ls()))
+  return(q_T)
+}
+
 cumDens <- function(x, probFunction, ...) {
   if (x == -Inf || x == 0) {
     return(0)