+++ tests/testthat/test-direct-sampling.R
@@ -153,6 +153,40 @@ test_that("Original attributes are retrieved", {
     c("shape" = 4, "rate" = 1 / 3),
     tolerance= 1e-1
   )
+
+  # Inv Gauss
+  invgauss <- rtrunc(1e5, m = 61, s = 7, faster = TRUE, family = "invgauss")
+  expect_equal(
+    attributes(invgauss),
+    list(
+      "class" = "trunc_invgauss",
+      "parameters" = list("m" = 61, "s" = 7),
+      "truncation_limits" = list("a" = 0, "b" = Inf),
+      "continuous" = TRUE
+    )
+  )
+  expect_equal(
+    mlEstimationTruncDist(invgauss, delta = .01),
+    c("m" = 61, "s" = 7),
+    tolerance= 1e-1
+  )
+
+  # Log-normal
+  invlnorm <- rtrunc(1e5, meanlog = 4, sdlog = 1, faster = TRUE, family = "lognormal")
+  expect_equal(
+    attributes(invlnorm),
+    list(
+      "class" = "trunc_lognormal",
+      "parameters" = list("meanlog" = 4, "sdlog" = 1),
+      "truncation_limits" = list("a" = 0, "b" = Inf),
+      "continuous" = TRUE
+    )
+  )
+  expect_equal(
+    mlEstimationTruncDist(invlnorm),
+    c("meanlog" = 4, "sdlog" = 1),
+    tolerance= 1e-1
+  )
 })
 
 Sys.setenv("LANGUAGE" = "en")
@@ -227,6 +261,45 @@ test_that("Truncation is not a speed limiter", {
   )
   expect_length(
     rtrunc(n, family = "gamma", shape = 5, rate = 4, a = 4, faster = TRUE),
+    n
+  )
+
+  # Inv Gamma
+  expect_error({
+      setTimeLimit(time_limit)
+      summary(rtrunc(n, family = "invgamma", shape = 5, rate = 4, a = 4))
+    },
+    "reached CPU time limit"
+  )
+  expect_length(
+    rtrunc(n, family = "invgamma", shape = 5, rate = 4, a = 4, faster = TRUE),
+    n
+  )
+
+  # Inv Gauss
+  expect_error({
+      setTimeLimit(time_limit)
+      summary(rtrunc(n, family = "invgauss", m = 5, s = 4, a = 4, b = 100))
+    },
+    "reached CPU time limit"
+  )
+  expect_length(
+    rtrunc(n, family = "invgauss", m = 5, s = 4, a = 4, b = 100, faster = TRUE),
+    n
+  )
+
+  # Lognormal
+  expect_error({
+      setTimeLimit(time_limit)
+      rtrunc(n, family = "lognormal", meanlog = 5, sdlog = 2, a = 150, b = 160)
+    },
+    "reached CPU time limit"
+  )
+  expect_length(
+    rtrunc(
+      n, family = "lognormal", meanlog = 5, sdlog = 2, a = 150, b = 160,
+      faster = TRUE
+    ),
     n
   )
 })